refactor(projects): split project ops hook
